+++ lib/gpio.ex
@@ -23,10 +23,31 @@ defmodule GpioRpi do
   Start and link a new GPIO GenServer. `pin` should be a valid
   GPIO pin number on the system and `pin_direction` should be
   `:input` or `:output`.
+  Optional parameters:
+  `:mode` set pullup register, see set_mode/2.
+  `:interrupt`, enable interrupts, see set_int/2 for the different transitions.
   """
   @spec start_link(integer, pin_direction, [term]) :: {:ok, pid}
   def start_link(pin, pin_direction, opts \\ []) do
-    GenServer.start_link(__MODULE__, [pin, pin_direction], opts)
+    {pullup, opts} = Keyword.pop(opts, :mode, :unset)
+    {interrupt, opts} = Keyword.pop(opts, :interrupt, :unset)
+
+    {:ok, pid} = GenServer.start_link(__MODULE__, [pin, pin_direction], opts)
+
+    case pin_direction do
+      :input ->
+        case pullup do
+          :unset -> :ok
+          mode -> set_mode(pid, mode)
+        end
+        case interrupt do
+          :unset -> :ok
+          dir -> set_int(pid, dir)
+        end
+      _ -> :ok
+    end
+
+    {:ok, pid}
   end
 
   @doc """
@@ -77,6 +98,30 @@ defmodule GpioRpi do
     GenServer.call pid, {:set_mode, pullup_mode}
   end
 
+  @doc """
+  Set the input/output direction of the pin.
+  Additionally allows setting `:mode` and `:interrupt'
+  """
+  @spec set_direction(pid, pin_direction, [term]) :: :ok | {:error, term}
+  def set_direction(pid, pin_direction, opts \\ []) do
+
+    :ok = GenServer.call pid, {:set_direction, pin_direction}
+
+    case pin_direction do
+      :input ->
+        case Keyword.get(opts, :mode, :unset) do
+          :unset -> :ok
+          mode -> set_mode(pid, mode)
+        end
+        case Keyword.get(opts, :interrupt, :unset) do
+          :unset -> :ok
+          dir -> set_int(pid, dir)
+        end
+      _ -> :ok
+    end
+    :ok
+  end
+
   # gen_server callbacks
   def init([pin, pin_direction]) do
     executable = :code.priv_dir(:gpio_rpi) ++ '/gpio_rpi'
@@ -104,8 +149,13 @@ defmodule GpioRpi do
     state = %{state | callbacks: new_callbacks}
     {:reply, response, state}
   end
+
   def handle_call({:set_mode, mode}, _from, state) do
     {:ok, response} = call_port(state, :set_mode, mode)
+    {:reply, response, state}
+  end
+  def handle_call({:set_direction, pin_direction}, _from, state) do
+    {:ok, response} = call_port(state, :set_direction, pin_direction)
     {:reply, response, state}
   end