+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.2.0
+  * Allow setting of pullup on initialization of pin.
+  * Allow changing the direction, mode, interrupt of a pin.
+
 ## v0.1.0
   * Extract of GPIO part of elixir_ale.
     * In agreement with fhunleth, maintainer of elixir_ale, we decided to make a seperate